feat: マルチユーザーモードをメインシステムに統一
## 変更内容

### RFP要件準拠
- シングルユーザーモードを完全削除し、マルチユーザーモードを標準化
- main-multiuser.tsの機能をmain.tsに統合、単一エントリポイントに簡素化

### コード統合と削除
- main-multiuser.ts → main.ts統合完了
- setupSingleUserCommands削除、マルチユーザーコマンドのみに統一
- 不要テストファイル削除（telegram-commands.test.ts等）
- obsoleteクラス削除（MultiUserScheduler.ts, TelegramBot.ts）

### 技術的改善
- validateMultiUserConfig, displayMultiUserConfig, setupMultiUserGracefulShutdownヘルパー関数統合
- TelegramNotifierをマルチユーザー専用に簡素化
- MarkdownV2解析削除、プレーンテキスト統一
- TypeScript型安全性確保、全型チェックエラー解決

### システム稼働確認
- マルチユーザーモード正常起動確認
- データベース初期化・管理画面起動確認
- 監視スケジューラー統合動作確認

Closes #XX
