+++ src/main.ts
@@ -1,97 +1,117 @@
 import { config, validateConfig, displayConfig } from './config.js';
-import { MonitoringScheduler } from './scheduler.js';
+import { MultiUserMonitoringScheduler } from './scheduler.js';
+import { UserService } from './services/UserService.js';
+import { TelegramNotifier } from './telegram.js';
 import { logger, vibeLogger } from './logger.js';
 import { performanceMonitor } from './performance.js';
+import { AppDataSource } from './database/connection.js';
 
 /**
  * メイン関数
  */
 async function main(): Promise<void> {
-  // マルチユーザーモードチェック
-  if (config.multiUser?.enabled) {
-    console.log('🔄 マルチユーザーモードで起動します...');
-    const { startMultiUserMode } = await import('./main-multiuser.js');
-    await startMultiUserMode();
-    return;
-  }
-
   console.log('===========================================');
-  console.log('   ソクブツ MVP - 新着物件通知サービス   ');
-  console.log('     完全リセット戦略準拠実装版         ');
+  console.log('   ソクブツ MVP - マルチユーザーモード    ');
+  console.log('    Puppeteer-first戦略＋複数URL対応     ');
   console.log('===========================================');
   console.log(`起動時刻: ${new Date().toLocaleString('ja-JP', { timeZone: 'Asia/Tokyo' })}`);
   console.log();
 
   // パフォーマンス監視開始
-  vibeLogger.info('main.startup', 'パフォーマンス監視開始', {
+  vibeLogger.info('multiuser.main.startup', 'マルチユーザーモード起動開始', {
     context: {
       startupTime: new Date().toISOString(),
       nodeVersion: process.version,
       platform: process.platform,
+      mode: 'multiuser',
     },
-    humanNote: 'アプリケーションの起動プロセスを開始',
+    humanNote: 'マルチユーザーモードでアプリケーションを起動',
   });
 
   // 設定検証
-  if (!validateConfig()) {
-    console.error('\\n🚨 環境変数を .env ファイルに設定してください。');
-    console.error('\\n📝 設定例:');
+  if (!validateMultiUserConfig()) {
+    console.error('\n🚨 マルチユーザーモード用の環境変数を設定してください。');
+    console.error('\n📝 設定例:');
+    console.error('MULTI_USER_MODE=true');
     console.error('TELEGRAM_BOT_TOKEN=your_bot_token_here');
-    console.error('TELEGRAM_CHAT_ID=your_chat_id_here');
-    console.error('MONITORING_URLS=https://www.athome.co.jp/chintai/tokyo/shinjuku-city/list/');
-    console.error('\\n💡 詳細は README.md をご確認ください。');
+    console.error('DATABASE_URL=./data/sokubutsu.db');
+    console.error('\n💡 詳細は README.md をご確認ください。');
     process.exit(1);
   }
 
   // 設定情報表示
-  displayConfig();
+  displayMultiUserConfig();
   console.log();
 
   // パフォーマンス指標表示
   performanceMonitor.displayMetrics();
   console.log();
 
-  // 管理画面サーバー起動（Admin機能が有効な場合）
-  if (config.admin?.enabled) {
-    const { AdminServer } = await import('./admin/AdminServer.js');
-    const adminServer = new AdminServer();
-    adminServer.start(config.admin.port || 3001);
-    console.log(`\\n📊 管理画面が起動しました: http://localhost:${config.admin.port || 3001}`);
+  // データベース初期化
+  try {
+    await AppDataSource.initialize();
+    vibeLogger.info('multiuser.db.initialized', 'データベース初期化完了', {
+      context: { databasePath: config.database?.database || './data/sokubutsu.db', entities: AppDataSource.entityMetadatas.length },
+      humanNote: 'データベースが正常に初期化されました',
+    });
+  } catch (error) {
+    vibeLogger.error('multiuser.db.initialization_error', 'データベース初期化エラー', {
+      context: { error: error instanceof Error ? error.message : String(error) },
+    });
+    console.error('🚨 データベース初期化エラー:', error);
+    process.exit(1);
   }
 
-  // スケジューラー起動
-  const scheduler = new MonitoringScheduler(config.telegram.botToken, config.telegram.chatId);
+  // マイグレーション実行
+  await AppDataSource.runMigrations();
+  vibeLogger.info('multiuser.db.migrations_complete', 'データベースマイグレーション完了', {
+    humanNote: 'データベーススキーマが最新版に更新されました',
+  });
+
+  console.log('✅ データベース接続確認完了');
+  console.log();
+
+  // 管理画面サーバー起動（マルチユーザーモードでは必須）
+  const { AdminServer } = await import('./admin/AdminServer.js');
+  const adminServer = new AdminServer();
+  adminServer.start(config.admin?.port || 3001);
+  console.log(`\n📊 管理画面が起動しました: http://localhost:${config.admin?.port || 3001}`);
+
+  // マルチユーザー監視スケジューラー起動
+  const scheduler = new MultiUserMonitoringScheduler(config.telegram.botToken);
+  const userService = scheduler.getUserService();
 
   try {
-    await scheduler.start(config.monitoring.urls, config.telegram.enabled);
-    
-    // Telegramコマンドハンドラーを設定（シングルユーザーモード）
-    const telegram = scheduler.getTelegram();
-    telegram.setupCommandHandlers(scheduler);
+    // 先にTelegram Botを起動して、初回チェック中でもコマンドを受け付け可能にする
+    const telegram = new TelegramNotifier(config.telegram.botToken, config.telegram.chatId);
+    telegram.setupCommandHandlers(scheduler, userService);
     await telegram.launchBot();
 
-    console.log('✅ 監視を開始しました。5分間隔で実行されます。');
-    console.log('🤖 Telegram Botコマンドが利用可能です。');
-    console.log('📊 統計レポートは1時間ごとに送信されます。');
+    // 監視の起動（初回チェックなどは並行で進む）
+    await scheduler.start();
+
+    console.log('✅ マルチユーザー監視を開始しました。5分間隔で実行されます。');
+    console.log('🤖 Telegram Botマルチユーザーコマンドが利用可能です。');
+    console.log('📊 ユーザー別統計レポートは1時間ごとに送信されます。');
     console.log('🛑 停止するには Ctrl+C を押してください。');
     console.log();
 
-    vibeLogger.info('main.startup_complete', 'ソクブツMVP起動完了', {
+    vibeLogger.info('multiuser.main.startup_complete', 'マルチユーザーモード起動完了', {
       context: {
-        urlCount: config.monitoring.urls.length,
+        mode: 'multiuser',
         interval: config.monitoring.interval,
         performance: performanceMonitor.getMetrics(),
       },
-      humanNote: 'システムが正常に起動し、監視を開始しました',
+      humanNote: 'マルチユーザーシステムが正常に起動し、監視を開始しました',
     });
 
     // グレースフルシャットダウン設定
-    setupGracefulShutdown(scheduler);
+    setupMultiUserGracefulShutdown(scheduler, telegram);
 
     // プロセスを維持
     process.stdin.resume();
   } catch (error) {
-    vibeLogger.error('main.startup_error', '起動エラー', {
+    vibeLogger.error('multiuser.main.startup_error', 'マルチユーザーモード起動エラー', {
       context: {
         error:
           error instanceof Error
@@ -102,24 +122,63 @@ async function main(): Promise<void> {
               }
             : { message: String(error) },
       },
-      aiTodo: '起動エラーの原因を分析し、解決策を提案',
+      aiTodo: 'マルチユーザーモード起動エラーの原因を分析し、解決策を提案',
     });
-    console.error('🚨 起動エラー:', error instanceof Error ? error.message : error);
+    console.error('🚨 マルチユーザーモード起動エラー:', error instanceof Error ? error.message : error);
     process.exit(1);
   }
 }
 
 /**
- * グレースフルシャットダウン設定
+ * マルチユーザーモード設定検証
  */
-function setupGracefulShutdown(scheduler: MonitoringScheduler): void {
+function validateMultiUserConfig(): boolean {
+  const requiredVars = [
+    'TELEGRAM_BOT_TOKEN',
+  ];
+
+  for (const varName of requiredVars) {
+    if (!process.env[varName]) {
+      console.error(`❌ 環境変数 ${varName} が設定されていません`);
+      return false;
+    }
+  }
+
+  // マルチユーザーモードが有効かチェック
+  if (!config.multiUser?.enabled) {
+    console.error('❌ MULTI_USER_MODE=true が設定されていません');
+    return false;
+  }
+
+  return true;
+}
+
+/**
+ * マルチユーザーモード設定表示
+ */
+function displayMultiUserConfig(): void {
+  console.log('📋 マルチユーザーモード設定:');
+  console.log(`  • Telegram Bot Token: ${config.telegram.botToken ? '✅ 設定済み' : '❌ 未設定'}`);
+  console.log(`  • データベース: ${config.database?.database || './data/sokubutsu.db'}`);
+  console.log(`  • 管理画面: ${config.admin?.enabled ? '✅ 有効' : '❌ 無効'}`);
+  console.log(`  • 監視間隔: ${config.monitoring.interval || '*/5 * * * *'}`);
+  console.log(`  • サーキットブレーカー: ${config.circuitBreaker.autoRecoveryEnabled ? '✅ 自動復旧' : '⚠️ 手動復旧'}`);
+}
+
+/**
+ * マルチユーザーモードグレースフルシャットダウン設定
+ */
+function setupMultiUserGracefulShutdown(
+  scheduler: MultiUserMonitoringScheduler,
+  telegram: TelegramNotifier
+): void {
   const shutdown = (signal: string) => {
     vibeLogger.info(
-      'main.shutdown_signal',
-      `${signal} 信号を受信しました。シャットダウンを開始します。`,
+      'multiuser.main.shutdown_signal',
+      `${signal} 信号を受信しました。マルチユーザーモードをシャットダウンします。`,
       {
-        context: { signal },
-        humanNote: 'グレースフルシャットダウンを実行中',
+        context: { signal, mode: 'multiuser' },
+        humanNote: 'マルチユーザーモードでグレースフルシャットダウンを実行中',
       }
     );
     console.log(`\n\n🛑 ${signal} 信号を受信しました。`);
@@ -129,15 +188,28 @@ function setupGracefulShutdown(scheduler: MonitoringScheduler): void {
     performanceMonitor.displayMetrics();
 
     // Telegram Botを停止
-    const telegram = scheduler.getTelegram();
     telegram.stopBot();
 
-    // スケジューラー停止
+    // マルチユーザースケジューラー停止
     scheduler.stop();
 
-    console.log('✅ ソクブツMVPを正常に終了しました。');
-    vibeLogger.info('main.shutdown_complete', 'ソクブツMVP正常終了', {
-      humanNote: 'アプリケーションが正常に終了しました',
+    // データベース接続を閉じる
+    if (AppDataSource.isInitialized) {
+      AppDataSource.destroy()
+        .then(() => {
+          vibeLogger.info('multiuser.db.closed', 'データベース接続を閉じました');
+        })
+        .catch((error) => {
+          vibeLogger.error('multiuser.db.close_error', 'データベース接続終了エラー', {
+            context: { error: error instanceof Error ? error.message : String(error) },
+          });
+        });
+    }
+
+    console.log('✅ マルチユーザーモードを正常に終了しました。');
+    vibeLogger.info('multiuser.main.shutdown_complete', 'マルチユーザーモード正常終了', {
+      context: { mode: 'multiuser' },
+      humanNote: 'マルチユーザーアプリケーションが正常に終了しました',
     });
 
     process.exit(0);
